refactor(lsp): add lsp mod
