+++ packages/ds-pinyin-lsp/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod lsp;
 pub mod sqlite;
 pub mod types;
 pub mod utils;
